backend/app/requirements.txt
fastapi
uvicorn
xarray
cfgrib
matplotlib
pillow
requests
beautifulsoup4